+++ common/src/main/java/com/gentics/mesh/storage/AbstractBinaryStorage.java
@@ -1,4 +1,4 @@
-package com.gentics.mesh.storage;
+package com.gentics.mesh.core.data.storage;
 
 /**
  * @see BinaryStorage
