+++ core/src/main/java/com/gentics/mesh/core/endpoint/node/S3BinaryUploadHandlerImpl.java
@@ -38,7 +38,7 @@ import com.gentics.mesh.core.rest.schema.S3BinaryFieldSchema;
 import com.gentics.mesh.core.verticle.handler.HandlerUtilities;
 import com.gentics.mesh.etc.config.MeshOptions;
 import com.gentics.mesh.json.JsonUtil;
-import com.gentics.mesh.storage.S3BinaryStorage;
+import com.gentics.mesh.core.data.storage.S3BinaryStorage;
 import com.gentics.mesh.util.UUIDUtil;
 
 import io.reactivex.Observable;
