+++ mdm/api/src/main/java/com/gentics/mesh/core/data/storage/LocalBinaryStorage.java
@@ -1,4 +1,4 @@
-package com.gentics.mesh.storage;
+package com.gentics.mesh.core.data.storage;
 
 /**
  * Storage definition for local filesystem binary data storages.
