+++ mdm/common/src/main/java/com/gentics/mesh/core/data/dao/PersistingBinaryDao.java
@@ -1,12 +1,17 @@
 package com.gentics.mesh.core.data.dao;
 
 import java.io.InputStream;
+import java.util.Base64;
 import java.util.stream.Stream;
 
 import com.gentics.mesh.core.data.binary.Binaries;
 import com.gentics.mesh.core.data.binary.HibBinary;
+import com.gentics.mesh.core.data.storage.BinaryStorage;
 import com.gentics.mesh.core.db.Supplier;
 import com.gentics.mesh.core.db.Transactional;
+import com.gentics.mesh.core.db.Tx;
+import io.reactivex.Flowable;
+import io.vertx.core.buffer.Buffer;
 
 /**
  * Persistence-aware extension to {@link BinaryDao}
@@ -15,6 +20,8 @@ import com.gentics.mesh.core.db.Transactional;
  *
  */
 public interface PersistingBinaryDao extends BinaryDao {
+
+	Base64.Encoder BASE64 = Base64.getEncoder();
 
 	/**
 	 * Get a binary storage implementation.
@@ -41,5 +48,17 @@ public interface PersistingBinaryDao extends BinaryDao {
 	@Override
 	default Supplier<InputStream> openBlockingStream(HibBinary binary) {
 		return binary.openBlockingStream();
+	}
+
+	@Override
+	default Flowable<Buffer> getStream(HibBinary binary) {
+		BinaryStorage storage = Tx.get().data().binaryStorage();
+		return storage.read(binary.getUuid());
+	}
+
+	@Override
+	default String getBase64ContentSync(HibBinary binary) {
+		Buffer buffer = Tx.get().data().binaryStorage().readAllSync(binary.getUuid());
+		return BASE64.encodeToString(buffer.getBytes());
 	}
 }
