+++ mdm/orientdb-wrapper/src/main/java/com/gentics/mesh/core/data/s3binary/impl/S3BinaryImpl.java
@@ -14,7 +14,7 @@ import com.gentics.mesh.core.rest.MeshEvent;
 import com.gentics.mesh.core.rest.event.s3binary.S3BinaryEventModel;
 import com.gentics.mesh.core.result.Result;
 import com.gentics.mesh.madl.field.FieldType;
-import com.gentics.mesh.storage.S3BinaryStorage;
+import com.gentics.mesh.core.data.storage.S3BinaryStorage;
 
 /**
  * @see S3Binary
