+++ mdm/orientdb-wrapper/src/main/java/com/gentics/mesh/graphdb/tx/impl/TxDataImpl.java
@@ -6,6 +6,7 @@ import com.gentics.mesh.cli.OrientDBBootstrapInitializer;
 import com.gentics.mesh.core.data.HibMeshVersion;
 import com.gentics.mesh.core.data.dao.PermissionRoots;
 import com.gentics.mesh.core.data.service.ServerSchemaStorage;
+import com.gentics.mesh.core.data.storage.BinaryStorage;
 import com.gentics.mesh.core.db.CommonTxData;
 import com.gentics.mesh.core.db.TxData;
 import com.gentics.mesh.dagger.OrientDBMeshComponent;
@@ -58,5 +59,10 @@ public class TxDataImpl implements CommonTxData {
 	@Override
 	public ServerSchemaStorage serverSchemaStorage() {
 		return mesh().serverSchemaStorage();
+	}
+
+	@Override
+	public BinaryStorage binaryStorage() {
+		return mesh().binaryStorage();
 	}
 }
