+++ tests/tests-core/src/main/java/com/gentics/mesh/core/field/binary/BinaryFieldTest.java
@@ -182,7 +182,7 @@ public class BinaryFieldTest extends AbstractFieldTest<BinaryFieldSchema> {
 			field.cloneTo(otherContainer);
 
 			HibBinaryField clonedField = otherContainer.getBinary(BINARY_FIELD);
-			assertThat(clonedField).as("cloned field").isNotNull().isEqualToIgnoringGivenFields(field, "outV", "id", "uuid", "element");
+			assertThat(clonedField).as("cloned field").isNotNull().isEqualToIgnoringGivenFields(field, "outV", "id", "uuid", "element", "contentUUID", "dbUuid");
 			assertThat(clonedField.getBinary()).as("referenced binary of cloned field").isNotNull().isEqualToComparingFieldByField(field.getBinary());
 		}
 	}
