+++ tests/tests-core/src/main/java/com/gentics/mesh/core/schema/field/BinaryFieldMigrationTest.java
@@ -28,7 +28,7 @@ import com.gentics.mesh.core.data.node.field.HibBinaryField;
 import com.gentics.mesh.core.db.Tx;
 import com.gentics.mesh.core.field.DataProvider;
 import com.gentics.mesh.core.field.binary.BinaryFieldTestHelper;
-import com.gentics.mesh.storage.BinaryStorage;
+import com.gentics.mesh.core.data.storage.BinaryStorage;
 import com.gentics.mesh.test.MeshTestSetting;
 import com.gentics.mesh.util.FileUtils;
 import com.gentics.mesh.util.RxUtil;
