+++ tests/tests-service-local-storage/src/main/java/com/gentics/mesh/storage/LocalBinaryStorageTest.java
@@ -1,4 +1,4 @@
-package com.gentics.mesh.storage;
+package com.gentics.mesh.core.data.storage;
 
 public class LocalBinaryStorageTest {
